requirements/base.txt
beautifulsoup4==4.6.0
django-environ==0.4.4
django-model-utils==3.1.1
Django==2.0.3
djangorestframework==3.7.7
requests==2.18.4
tqdm==4.19.9
